feat: refactor GraphRequest, Transactions value type to double and interface DI

+++ software/core/oqm-core-base-station/src/main/java/tech/ebp/oqm/core/baseStation/model/graph/Transactions.java
@@ -2,4 +2,4 @@ package tech.ebp.oqm.core.baseStation.model.graph;
 
 import java.time.Instant;
 
-public record Transactions(Instant timestamp, int value) {}
+public record Transactions(Instant timestamp, double value) {}
